Added multisite support for uninstall.php
Added multisite support for uninstall.php

+++ uninstall.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Remove all options on uninstall 
+ * Remove all options on uninstall
  *
  * @since      1.0.0
  * @package    Polygon_Plugin
@@ -25,5 +25,23 @@ if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 /**
  * Remove plugin options.
  */
-delete_option( 'polygon_plugin' );
-delete_option( 'external_updates-polygon-plugin' );
+if ( is_multisite() ) {
+	// Global variables
+	global $wpdb;
+
+	// Variables
+	$blogs = $wpdb->get_results( "SELECT blog_id FROM {$wpdb->blogs}", ARRAY_A );
+
+	if ( $blogs ) {
+		foreach( $blogs as $blog ) {
+			switch_to_blog( $blog['blog_id'] );
+
+			delete_option( 'polygon_plugin' );
+			delete_option( 'external_updates-polygon-plugin' );
+		}
+		restore_current_blog();
+	}
+} else {
+	delete_option( 'polygon_plugin' );
+	delete_option( 'external_updates-polygon-plugin' );
+}
